Tests for feature rendering now invoke the CukedoctorFeatureRenderer rather than CukedoctorConverterImpl. Remove now unused renderFeatures method from CukedoctorConverter interface.
