+++ cukedoctor-converter/src/test/java/com/github/cukedoctor/bdd/cukedoctor/OrderingSteps.java
@@ -1,10 +1,8 @@
 package com.github.cukedoctor.bdd.cukedoctor;
 
 import com.github.cukedoctor.Cukedoctor;
-import com.github.cukedoctor.api.CukedoctorDocumentBuilder;
 import com.github.cukedoctor.api.model.Feature;
 import com.github.cukedoctor.builder.CukedoctorDocumentBuilderImpl;
-import com.github.cukedoctor.config.CukedoctorConfig;
 import com.github.cukedoctor.parser.FeatureParser;
 import com.github.cukedoctor.renderer.CukedoctorFeatureRenderer;
 import io.cucumber.java.en.Given;
@@ -68,7 +66,7 @@ public class OrderingSteps {
         assertThat(featureFile).isNotNull();
         List<Feature> features = FeatureParser.parse(featureFile.getPath());
         assertThat(features).isNotNull().hasSize(2);
-        documentation = Cukedoctor.instance(features, null, new CukedoctorConfig(), CukedoctorDocumentBuilder.Factory.newInstance().nestTitle().createNestedBuilder()).renderFeatures(features).getDocumentation();
+        documentation = Cukedoctor.instance(features).renderDocumentation();
     }
 
     private void assertFeaturesShouldBeOrdered(String expected) {
